Add files via upload
Lab 2
1. Compute the control digit of an integer by summing up its digits, then summing 
up the digits of the sum, so on, until a sum of only one digit is obtained.
e.g. The control digit of integer number 1971 is 9 (1971 → 18 → 9).

5. Determine the value of the element at index 𝒌 in the array 1, 2, 2, 3, 3, 3, 4, 4, 
4, 4,… without reading or effectively creating the array

9.Print all numbers with maximum 2 digits of form 𝒙𝒚 with the property that the 
last digit of (𝒙𝒚)^2 is 𝒚.
e.g. 5^2=25 or 10^2=100 or 76^2=5776

13. Read a natural number 𝒏. Form another number from its digits found at odd 
positions (from left to right).

17. A number 𝒏 is special if there is a natural number 𝒎 such that 𝑛 = 𝑚 + 𝑆(𝑚)
where 𝑆(𝑚) is the sum of digits of 𝒎. Verify if a given number is special.
e.g. 1235 is special (1235=1225+10)